+++ packages/rps/src/redux/auto-player/actions.ts
