+++ packages/rps/src/redux/auto-player/auto-opponent.ts
