+++ packages/rps/src/redux/auto-player/auto-player.ts
