+++ packages/rps/src/redux/auto-player/index.ts
