+++ packages/rps/src/redux/auto-player/reducer.ts
