+++ packages/rps/src/redux/auto-player/state.ts
